+++ demo/pages/search_page/components/package_list.py
@@ -8,27 +8,21 @@ from .package import Package
 class PackageList(ListComponent[Package, PyPIPage]):
     """Represent the list of search results on `SearchPage`."""
 
-    # The `ListComponent` item should always be `ComponentWithBaseLocator`,
-    # because all its methods depend on `base_locator`
+    # The ``ListComponent`` item should always be ``ComponentWithBaseLocator``,
+    # because all its methods depend on `base_locator`. Also this attribute is
+    # required.
     item_class = Package
 
-    def __init__(
-        self,
-        page: PyPIPage,
-        # We use the empty init to set the default value for `base_locator`
-        base_locator: locators.XPathLocator = locators.PropertyLocator(
-            prop="aria-label",
-            value="Search results",
-        ),
-        wait_until_visible: bool = True,
-    ):
-        super().__init__(page, base_locator, wait_until_visible)
+    base_locator = locators.PropertyLocator(
+        prop="aria-label",
+        value="Search results",
+    )
 
-    # Set up `base_item_locator` and `item_class` is required
-    @property
-    def base_item_locator(self) -> locators.XPathLocator:
-        """Get the base locator of result item."""
-        return self.base_locator // locators.ClassLocator(
-            class_name="package-snippet",
-            container="a",
-        )
+    # Set up ``relative_item_locator`` or ``item_locator`` is required.
+    # Use ``relative_item_locator`` - if you want locator nested within
+    # ``base_locator``, ``item_locator`` - otherwise."
+    # You also may override ``base_item_locator`` property.
+    relative_item_locator = locators.ClassLocator(
+        class_name="package-snippet",
+        container="a",
+    )